Added fulls screen dialog to AlertDialog.kt that shows up when creating a calender on the CreateFileScreen.kt.
Added date picker api to gradle to select dates and time when creating events/calender.

Implemented CreateFileScreen.kt. Added a floating action button that opens dialog to create a events for a calender with save and cancel options.

Updated event to have val keywords as they should not be changed once initialised.

Added skeleton code to ICSReader.kt for creating a calender.

Updated MainViewModel.kt to have a created events list with function that updates the list as the events are created.

Updated ReadFileScreen.kt to show a text when no file or url has been read and events are not shown.