Added coil dependency to gradle for showing gif.
Added a static and a gif logo to drawable.

Added a SplashScreen.kt to Screens.kt. Customised it to show an animated gif using coil.

Updated Navigation.kt to include SplashScreen.kt and made it the start destination.

Added a a function to MainViewModel.kt to simulate a asynchronous operation to keep the Splash screen showing.

Updated the HomeScreen.kt to show the animating logo. Made it interactive to toggle the animation, with visibility animations.

Updated EzCalenderApp.kt to only show the BottomNavigation.kt once after the spalsh screen.

Reformatted code/optimised imports.